+++ finRL/results/actions.csv
@@ -1,0 +1,124 @@
+AAPL,AMGN,AXP,BA,CAT,CRM,CSCO,CVX,DIS,GS,HD,HON,IBM,INTC,JNJ,JPM,KO,MCD,MMM,MRK,MSFT,NKE,PG,TRV,UNH,V,VZ,WBA,WMT
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,0,0,0,0,0,0,0,0,36,4,0,39,0,22,32,12,0,0,0,19,0,29,0,0,7,0,0,0
+0,13,3,0,0,0,0,5,0,35,0,0,31,0,17,36,12,0,0,0,17,0,37,0,0,9,0,0,0
+0,10,4,0,0,0,0,2,0,31,6,0,27,0,14,33,13,0,4,0,20,0,40,0,0,6,0,0,0
+0,0,0,0,0,0,0,3,0,0,-6,0,-14,1,-17,0,0,0,-4,0,20,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,0,0,0,-14,1,-22,0,0,0,0,0,15,0,-1,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-3,0,-8,2,-15,0,0,0,0,0,13,0,-2,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-3,0,-8,0,-15,0,0,0,0,0,13,0,-2,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-4,0,-12,2,-18,0,0,0,0,0,18,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-4,0,-12,0,-18,0,0,0,0,0,18,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,2,0,0,-4,0,-12,0,-18,0,0,0,0,0,17,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-4,0,-12,1,-18,0,0,0,0,0,18,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-4,0,-12,0,-18,0,0,0,0,0,18,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-4,0,-12,1,-18,0,0,0,0,0,18,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,2,0,0,-4,0,-9,0,-16,0,0,0,0,0,15,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,0,-4,0,-9,0,-16,0,0,0,0,0,15,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,0,-4,0,-9,0,-16,0,0,0,0,0,15,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-4,0,-12,1,-18,0,0,0,0,0,18,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,-4,0,-12,1,-18,0,0,0,0,0,18,0,-5,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,-10,0,-21,0,0,0,0,0,14,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,-10,0,-21,0,0,0,0,0,14,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,-10,0,-21,0,0,0,0,0,14,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,-10,1,-21,0,0,0,0,0,14,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,-10,1,-21,0,0,0,0,0,14,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,-10,0,-21,0,0,0,0,0,14,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-1,-1,0,-18,0,-25,0,0,0,0,0,20,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-6,0,0,-11,4,-18,0,0,0,0,0,18,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-6,0,0,-11,1,-18,0,0,0,0,0,19,0,-1,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-5,0,0,-11,0,-18,0,0,0,0,0,17,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-2,-1,0,-18,0,-26,0,0,0,0,0,21,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-2,0,0,-11,2,-19,0,0,0,0,0,14,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-2,0,0,-11,1,-19,0,0,0,0,0,14,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-2,0,0,-11,0,-7,0,0,0,0,0,9,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-2,0,0,-11,2,0,0,0,0,0,0,6,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-2,0,0,-11,2,0,0,0,0,0,0,6,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-4,0,0,-11,0,0,0,0,0,0,0,8,0,0,0,0,0,1,0,0
+0,0,0,0,0,0,0,1,0,-2,0,0,-11,1,0,0,0,0,0,0,6,0,0,0,0,0,-1,0,0
+0,1,0,0,0,0,0,0,0,0,0,0,-11,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,-11,5,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-1,-1,0,-18,0,0,0,0,0,0,0,9,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-1,-1,0,-18,0,0,0,0,0,0,0,9,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-1,-1,0,-18,1,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-1,-1,0,-18,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-4,-1,0,-17,1,0,-1,0,0,0,0,12,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,1,0,-1,-1,0,-18,1,0,0,0,0,0,0,9,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-1,-1,0,-18,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-1,-1,0,-18,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-1,-1,0,-18,3,0,0,0,0,0,0,9,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-3,0,0,-17,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-3,0,0,-17,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-1,-1,0,-18,1,0,0,0,0,0,0,9,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,2,0,0,0,0,-13,1,0,0,0,0,0,0,5,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,0,0,0,-17,1,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0
+0,0,-1,0,0,0,0,1,0,-1,0,0,-11,0,0,0,0,0,0,0,6,0,0,0,0,0,0,0,0
+0,6,0,0,0,0,0,18,0,-5,0,0,-14,0,0,0,-1,0,0,0,0,0,0,0,0,0,1,0,1
+0,6,0,1,0,0,0,23,0,-6,0,0,-16,0,0,0,-4,0,0,0,0,0,0,0,0,0,0,0,1
+0,0,0,0,0,0,0,0,0,-4,0,0,-14,3,0,0,-1,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-4,0,0,-14,3,0,0,-1,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-4,0,0,-14,2,0,0,-1,0,0,0,10,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-6,0,0,-14,0,0,0,0,0,0,0,12,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,1,0,-6,0,0,-14,0,0,0,0,0,0,0,11,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,1,0,-6,0,0,-14,0,0,0,0,0,0,0,11,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,1,0,-4,0,0,-13,0,0,0,0,0,0,0,10,0,-3,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-6,0,0,-14,0,0,0,0,0,0,0,12,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-5,0,0,-14,0,0,0,0,0,0,0,11,0,-1,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-6,0,0,-14,3,0,0,0,0,0,0,11,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-2,0,0,-12,2,0,0,-2,0,0,0,7,0,0,0,0,0,0,0,0
+0,0,-2,0,0,0,0,2,0,-3,0,0,-13,0,0,0,0,0,0,0,9,0,0,0,0,0,0,0,1
+0,0,-1,0,0,0,0,1,0,-3,0,0,-13,0,0,0,0,0,0,0,9,0,0,0,0,0,0,0,1
+0,0,-2,0,0,0,0,0,0,-3,0,0,-13,0,0,0,0,0,0,0,10,0,0,0,0,0,0,0,2
+0,0,-1,0,0,0,0,1,0,-3,0,0,-13,0,0,0,0,0,0,0,9,0,0,0,0,0,0,0,1
+0,0,0,0,0,0,0,0,0,-3,0,0,-13,0,0,0,0,0,0,0,9,0,0,0,0,0,0,0,1
+0,0,0,0,0,0,0,14,0,-6,0,0,-8,2,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-4,0,0,-11,0,0,0,0,0,0,0,9,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,17,0,-4,0,0,-11,16,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,18,0,-4,0,0,-11,17,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0
+0,0,3,0,0,0,0,17,0,-5,0,0,-8,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0
+0,0,0,0,0,0,0,20,0,-3,0,0,-11,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,0,0,-3,0,0,0,3,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,1,0,-3,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0
+0,0,0,2,0,0,0,0,0,-2,-1,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0
+0,0,0,2,0,0,0,0,0,-1,-1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,-1,0,0
+0,0,0,2,0,0,0,0,0,-1,-1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0
+0,0,0,2,0,0,0,0,0,-1,-1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0
+0,0,0,2,0,0,0,0,0,-3,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-3,0,0,0,0,0,0,0,0,0,0,3,0,0,0,0,0,0,0,0
+0,0,0,1,0,0,0,1,0,-1,-1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,-4,0,0,0,20,0,0,0,0,0,0,3,0,-3,0,0,0,0,0,0
+0,5,0,0,0,0,0,1,0,-3,0,0,0,0,0,0,0,0,0,0,0,0,-2,0,0,0,0,0,0
+0,5,0,0,0,0,0,1,0,-3,0,0,0,1,0,0,0,0,0,0,0,0,-2,0,0,0,0,0,0
+0,5,0,0,0,0,0,0,0,-3,0,0,0,1,0,0,0,0,0,0,0,0,-1,0,0,0,0,0,0
+0,5,0,0,0,0,0,0,0,-3,0,0,0,1,0,0,0,0,0,0,0,0,-1,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,2,0,0,0,0,0,1,0,-2,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,3,0,0,0,0,0,0,0,-2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,1,0,-3,0,0,0,0,0,0,-1,0,0,0,3,0,0,0,0,0,0,0,-1
