+++ finRL/results/results.csv
@@ -1,0 +1,125 @@
+date,account_value
+2023-10-25,1000000.0
+2023-10-26,999860.7997524109
+2023-10-27,998629.170400055
+2023-10-30,1000799.3249701541
+2023-10-31,1002051.8678108373
+2023-11-01,1003653.9395230261
+2023-11-02,1007053.2781133581
+2023-11-03,1011541.0745626073
+2023-11-06,1011854.0382236177
+2023-11-07,1012304.5099378205
+2023-11-08,1012634.2675247194
+2023-11-09,1009389.2453648072
+2023-11-10,1016119.6813382109
+2023-11-13,1015561.7619966582
+2023-11-14,1025731.6054344939
+2023-11-15,1027479.0085642546
+2023-11-16,1031060.8291621093
+2023-11-17,1030355.9763921204
+2023-11-20,1034585.578967041
+2023-11-21,1029808.2984491425
+2023-11-22,1036623.430349701
+2023-11-24,1038201.7177410279
+2023-11-27,1036909.9955485535
+2023-11-28,1039851.3922360077
+2023-11-29,1043496.416193039
+2023-11-30,1052098.4319291727
+2023-12-01,1060424.1304740373
+2023-12-04,1059786.8092803003
+2023-12-05,1050895.5335550425
+2023-12-06,1045902.5074843484
+2023-12-07,1049477.0258487207
+2023-12-08,1058287.5618750115
+2023-12-11,1060617.5227214624
+2023-12-12,1066263.0850455933
+2023-12-13,1078419.316418442
+2023-12-14,1090397.8321277886
+2023-12-15,1090129.4427754937
+2023-12-18,1090617.2785578729
+2023-12-19,1097951.241317608
+2023-12-20,1085370.304703186
+2023-12-21,1092849.1824815485
+2023-12-22,1095497.3053445397
+2023-12-26,1098869.1601974298
+2023-12-27,1102104.1457881432
+2023-12-28,1105727.3955091324
+2023-12-29,1105998.2120358124
+2024-01-02,1106263.9544001007
+2024-01-03,1097917.9600629732
+2024-01-04,1098887.0150219423
+2024-01-05,1100701.8214903336
+2024-01-08,1111874.8324534988
+2024-01-09,1106895.8045670853
+2024-01-10,1113758.0201399918
+2024-01-11,1113694.3063467257
+2024-01-12,1117093.2497184754
+2024-01-16,1120407.131784279
+2024-01-17,1115813.4819388886
+2024-01-18,1120320.8894382515
+2024-01-19,1134736.2244854432
+2024-01-22,1136463.6720864717
+2024-01-23,1138008.767703907
+2024-01-24,1140012.4058310434
+2024-01-25,1158425.6108706247
+2024-01-26,1151014.11954863
+2024-01-29,1160590.5455397798
+2024-01-30,1169542.4854440652
+2024-01-31,1151194.6041977312
+2024-02-01,1161345.3391832963
+2024-02-02,1172000.8631204606
+2024-02-05,1160469.7555594903
+2024-02-06,1163002.8211558342
+2024-02-07,1174335.2512050897
+2024-02-08,1172454.47880962
+2024-02-09,1178714.2206198655
+2024-02-12,1180250.195495594
+2024-02-13,1155852.2539606057
+2024-02-14,1160897.300818268
+2024-02-15,1169750.2810045434
+2024-02-16,1165408.0974454079
+2024-02-20,1164010.768060421
+2024-02-21,1168185.8919079439
+2024-02-22,1185111.45453389
+2024-02-23,1185493.1948105013
+2024-02-26,1180044.8067609407
+2024-02-27,1179411.1069453584
+2024-02-28,1185548.7803391572
+2024-02-29,1189309.782474518
+2024-03-01,1190919.5045960618
+2024-03-04,1194861.3616847002
+2024-03-05,1179434.7092935296
+2024-03-06,1179267.6919671558
+2024-03-07,1186218.3811043473
+2024-03-08,1181634.811871052
+2024-03-11,1179881.8023640024
+2024-03-12,1197569.004187153
+2024-03-13,1202416.7665084116
+2024-03-14,1209323.0545037119
+2024-03-15,1197805.103985882
+2024-03-18,1199057.0149950983
+2024-03-19,1209597.0170397912
+2024-03-20,1222813.2716459122
+2024-03-21,1243408.3428013993
+2024-03-22,1234090.1456632768
+2024-03-25,1222611.0229324838
+2024-03-26,1222508.4060042615
+2024-03-27,1233816.9257719612
+2024-03-28,1235250.9121617624
+2024-04-01,1234469.5229835701
+2024-04-02,1227454.0223679696
+2024-04-03,1224656.1257536507
+2024-04-04,1213084.7268918343
+2024-04-05,1226908.2583588716
+2024-04-08,1228147.536391083
+2024-04-09,1229851.326099377
+2024-04-10,1216781.780704857
+2024-04-11,1219884.1353508532
+2024-04-12,1194728.1362810289
+2024-04-15,1190718.1751820527
+2024-04-16,1187579.3645644074
+2024-04-17,1189087.2129247591
+2024-04-18,1180435.250079403
+2024-04-19,1179549.9202575341
+2024-04-22,1196716.5640985072
+2024-04-23,1213317.7008393023
