+++ finRL/test.py
@@ -1,3 +1,4 @@
+import os
 import pandas as pd
 import matplotlib.pyplot as plt
 import argparse
@@ -8,16 +9,14 @@ from finrl.meta.env_stock_trading.env_stocktrading import StockTradingEnv
 from finrl.agents.stablebaselines3.models import DRLAgent
 from stable_baselines3 import PPO
 from finrl.main import check_and_make_directories
-from finrl.config import INDICATORS, TRAINED_MODEL_DIR
-from finrl.config import INDICATORS
+from finrl.config import INDICATORS, TRAINED_MODEL_DIR, RESULTS_DIR, DATA_SAVE_DIR
 from finrl.plot import backtest_stats
 
 # Contestants are welcome to split the data in their own way for model tuning
 TRADE_START_DATE = '2023-10-25'
 TRADE_END_DATE = '2024-04-24'
-TRAIN_FILE_PATH = 'train_data.csv'
-TRADE_FILE_PATH = 'trade_data.csv'
-
+TRAIN_FILE_PATH =  os.path.join(DATA_SAVE_DIR, 'train_data.csv')
+TRADE_FILE_PATH = os.path.join(DATA_SAVE_DIR, 'trade_data.csv')
 
 
 # PPO configs
@@ -55,7 +54,7 @@ if __name__ == '__main__':
     TRADE_START_DATE = args.start_date
     TRADE_END_DATE = args.end_date
     
-    processed_full = handle_diff(args.data_file, args.train_file)
+    processed_full = handle_diff(args.trade_file_path, args.train_file_path)
     trade = data_split(processed_full, TRADE_START_DATE, TRADE_END_DATE)
     
     stock_dimension = len(trade.tic.unique())
@@ -99,7 +98,7 @@ if __name__ == '__main__':
     plt.figure()
     
     df_result_ppo.plot()
-    plt.savefig("plot.png")
+    plt.savefig(os.path.join(RESULTS_DIR, 'plot.png'))
     
-    df_result_ppo.to_csv("results.csv", index=False)
-    df_actions_ppo.to_csv("actions.csv", index=False)
+    df_result_ppo.to_csv(os.path.join(RESULTS_DIR, 'results.csv'), index=False)
+    df_actions_ppo.to_csv(os.path.join(RESULTS_DIR, 'actions.csv'), index=False)
